+++ srcs/containers/django/src/pong/static/js/game/assets/Scene.js
@@ -1,11 +1,11 @@
-import * as THREE from '../utils/three.module.js';
+import * as THREE from '../THREE-utils/three.module.js';
 import { Ball } from './Ball.js';
 import { Field } from './Field.js';
 import { Paddle } from './Paddle.js';
 import { Camera } from './Camera.js';
 import { Audio } from './Audio.js';
 import { Environment } from './Environment.js';
-import { OrbitControls } from '../utils/OrbitControls.js';
+import { OrbitControls } from '../THREE-utils/OrbitControls.js';
 
 class Scene {
 	constructor(container) {
