+++ srcs/containers/django/src/pong/templates/main/base.html
@@ -12,9 +12,8 @@
       integrity="sha384-xOolHFLEh07PJGoPkLv1IbcEPTNtaed2xpHsD9ESMhqIYd0nLMwNLD69Npy4HI+N"
       crossorigin="anonymous"/>
     <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
-	  <script type="module" src="{% static 'js/game/app.js' %}" defer></script>
-    <!-- <script type="module" src="{% static 'js/module/2d-pong/entrypoint.js' %}" defer ></script> -->
-    <script type="module" src="{% static 'js/utils/page-loading.js' %}" defer></script>
+	  <script type="module" src="{% static 'js/game/update.js' %}" defer></script>
+    <script type="module" src="{% static 'js/page-loading/page-loading.js' %}" defer></script>
   </head>
 
   <body>
